[FIXED] compute shaders

+++ include/terraingraph/device/ThermalWeathering.parm.h
@@ -3,5 +3,5 @@
 #endif
 
 PARAM_INFO(Iterations,        int,   iter,                m_iter,                (1))
-PARAM_INFO(Amplitude,         int,   amplitude,           m_amplitude,           (1))
+PARAM_INFO(Amplitude,         float, amplitude,           m_amplitude,           (1.0))
 PARAM_INFO(TanThresholdAngle, float, tan_threshold_angle, m_tan_threshold_angle, (0.6))
